Update write_first_app.rst
Added a note at line 169 for users of Docker Toolbox, who will need to change the localhost to the docker IP for their local machine. I also added an answer on Stack Overflow here: https://stackoverflow.com/questions/43427448/something-wrong-on-deploy-chaincode-for-hyperledger-v1-0/47099908#47099908

Unrelated, but thought I'd mention it here, is that the function name in the node.js version of fabcar has a mismatched case from the documentation, which causes invoke.js to fail. In the docs, it's changeCarOwner (uppercase 'O'), and in node/fabcar.js, it's changeCarowner (lowercase 'o'). That would be a bug in the code, not in the documentation.

Thank you -- I'm enjoying the tutorial otherwise!